Moved all loggers and base logger to Loggers folder
